requirements.txt
olefile==0.44
Pillow==4.3.0
